requirements.txt
requests
fitz
flask
Pillow
pytesseract
